requirements.txt
altair
certifi
fastapi
Fortuna
Jinja2
joblib
jupyter
numpy
MonsterLab
pandas
motor
dnspython
python-dotenv
python-multipart
scikit-learn
scipy
uvicorn[standard]
